+++ frontend/src/components/shop/DisplayShops.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { connect } from "react-redux";
 import { Grid, IconButton, TextField } from "@material-ui/core";
-import { setShops } from "../redux/actions";
-import ShopService from "../services/ShopService";
+import { setShops } from "../../redux/actions";
+import ShopService from "../../services/ShopService";
 import Shop from "./Shop";
 import AddCircleRoundedIcon from "@material-ui/icons/AddCircleRounded";
 import AddShop from "./AddShop";
 import Select from "@material-ui/core/Select";
-import { PRIMARY_THEME_COLOR } from "../constants/constants";
+import { PRIMARY_THEME_COLOR } from "../../constants/constants";
 
 const mapStateToProps = (state) => {
   return { shops: state.shops };
