+++ frontend/src/components/shop/EditShop.js
@@ -10,49 +10,62 @@ import {
   InputAdornment,
   IconButton,
 } from "@material-ui/core";
+import Alert from "@material-ui/lab/Alert";
 import AddCircleIcon from "@material-ui/icons/AddCircle";
 import Shop from "./Shop";
-import { PRIMARY_THEME_COLOR } from "../constants/constants";
-import ShopService from "../services/ShopService";
+import { PRIMARY_THEME_COLOR } from "../../constants/constants";
+import ShopService from "../../services/ShopService";
 
-class AddShop extends React.Component {
+class EditShop extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
+      // shopData: this.props.selectedShop,
+      shopData: null,
       ownerEmail: "",
+      shopID: "",
       name: "",
       description: "",
       picture: "",
       tag: "",
       tags: [],
-      helper: "",
-      missingFields: false,
     };
     this.getShopData = this.getShopData.bind(this);
     this.deleteAllTags = this.deleteAllTags.bind(this);
+    this.updateShop = this.updateShop.bind(this);
+    this.deleteShop = this.deleteShop.bind(this);
+    this.discardChanges = this.discardChanges.bind(this);
     this.addTag = this.addTag.bind(this);
-    this.addShop = this.addShop.bind(this);
     this.mounted = true;
+  }
+
+  discardChanges() {
+    this.setState({
+      name: this.props.selectedShop.name,
+      description: this.props.selectedShop.description,
+      picture: this.props.selectedShop.picture,
+      tags: this.props.selectedShop.tags,
+    });
   }
 
   shouldComponentUpdate(nextProps, nextState) {
     return this.mounted;
   }
 
-  async addShop() {
-    if (this.state.name !== "" && this.state.description !== "") {
-      const success = await ShopService.createShop(this.getShopData());
-      if (success) {
-        this.props.onClose();
-      }
-    } else {
-      console.log("empty shop");
-      this.setState({ missingFields: true });
-    }
-  }
-
   componentWillUnmount() {
     this.mounted = false;
+  }
+
+  componentDidMount() {
+    this.setState({
+      shopData: this.props.selectedShop,
+      ownerEmail: this.props.selectedShop.ownerEmail,
+      shopID: this.props.selectedShop.shopID,
+      name: this.props.selectedShop.name,
+      description: this.props.selectedShop.description,
+      picture: this.props.selectedShop.picture,
+      tags: this.props.selectedShop.tags,
+    });
   }
 
   addTag() {
@@ -69,6 +82,7 @@ class AddShop extends React.Component {
 
   getShopData() {
     const shopObj = {
+      shopID: this.state.shopID,
       ownerEmail: this.state.ownerEmail,
       name: this.state.name,
       description: this.state.description,
@@ -79,15 +93,35 @@ class AddShop extends React.Component {
     return shopObj;
   }
 
+  async updateShop() {
+    const success = await ShopService.updateShop(this.getShopData());
+    if (success) {
+      this.props.onClose();
+    }
+  }
+
+  async deleteShop() {
+    const success = await ShopService.deleteShop(this.state.shopID);
+    if (success) {
+      this.props.onClose();
+    }
+  }
+
   render() {
     return (
       <Dialog
         fullWidth={true}
         maxWidth="md"
-        open={this.props.addingShop}
+        open={this.props.selectedShop !== null}
         onClose={this.props.onClose}
       >
-        <DialogTitle>Add Shop Menu</DialogTitle>
+        <DialogTitle>Edit Shop Menu</DialogTitle>
+
+        <DialogContent>
+          <Alert severity="error">
+            Deleting a shop will delete all the products it contains.
+          </Alert>
+        </DialogContent>
 
         <DialogContent style={{ display: "flex" }}>
           <div
@@ -110,10 +144,6 @@ class AddShop extends React.Component {
               fullWidth
               label="Update Name"
               value={this.state.name}
-              error={this.state.missingFields && this.state.name === ""}
-              helperText={
-                this.state.missingFields && this.state.name === "" && "Required"
-              }
               onChange={(event) => {
                 this.setState({ name: event.target.value });
               }}
@@ -127,12 +157,6 @@ class AddShop extends React.Component {
               rowsMax={4}
               multiline
               value={this.state.description}
-              error={this.state.missingFields && this.state.description === ""}
-              helperText={
-                this.state.missingFields &&
-                this.state.description === "" &&
-                "Required"
-              }
               onChange={(event) => {
                 this.setState({ description: event.target.value });
               }}
@@ -180,9 +204,19 @@ class AddShop extends React.Component {
 
         <DialogActions>
           <Chip
-            label="Create Shop"
-            onClick={this.addShop}
+            label="Update Shop"
+            onClick={this.updateShop}
             style={{ background: PRIMARY_THEME_COLOR, color: "white" }}
+          />
+          <Chip
+            label="Discard Changes"
+            onClick={this.discardChanges}
+            style={{ background: "#ff9800", color: "white" }}
+          />
+          <Chip
+            label="Delete Shop"
+            onClick={this.deleteShop}
+            color="secondary"
           />
         </DialogActions>
       </Dialog>
@@ -190,4 +224,4 @@ class AddShop extends React.Component {
   }
 }
 
-export default AddShop;
+export default EditShop;
